requirements.txt
streamlit
streamlit-player
python3-discogs-client
spotipy
sys
os
python-dotenv
pandas
matplotlib.pyplot
opencv-python-headless
PIL
google-cloud-storage
